output/36/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "docling-milvus-rag"
version = "0.1.0"
description = "Offline Docling → Milvus RAG pipeline packaged for uv."
authors = [{ name = "Task Machine", email = "ops@example.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.2.0",
    "pymilvus>=2.4.0",
    "sentence-transformers>=2.7.0",
    "langchain-text-splitters>=0.0.2",
    "structlog>=24.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "numpy>=1.26.0",
    "llama-cpp-python>=0.2.80",
    "pypdf>=4.0.1",
    "typer>=0.12.3",
    "rich>=13.7.0",
]

[project.scripts]
docling-milvus-rag = "docling_milvus_rag.cli:entrypoint"

[project.optional-dependencies]
dev = ["pytest>=8.2.0"]

[tool.uv]
dev-dependencies = ["pytest>=8.2.0"]
